+++ src/shared/utils/generate-random-code.ts
@@ -1,0 +1,14 @@
+const LENGTH_CODE = 6;
+
+export function generateRandomCode(): string {
+  const chars =
+    "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
+  let code = "";
+
+  for (let i = 0; i < LENGTH_CODE; i++) {
+    const randomIndex = Math.floor(Math.random() * chars.length);
+    code += chars[randomIndex];
+  }
+
+  return code;
+}